+++ HISTORY.md
@@ -2,9 +2,13 @@
 
 ## in development:
 
+* **Notable Changes:**
+    * added meaningful defaults to blocks in the palette that didn't already have them
 * **Notable Fixes:**
     * fixed a costume-shrinkWrap edgecase bug, thanks, Brian, for reporting it!
 
+### 2020-11-11
+* objects: added meaningful defaults to blocks in the palette that didn't already have them
 
 ### 2020-11-09
 * new dev version